pyproject.toml
[project]
name = "electricity-cost"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pandas>=2.3.1",
    "pymysql>=1.1.1",
    "python-dotenv>=1.1.1",
    "scikit-learn>=1.7.1",
]
